Refactoring to improve maintainability of `vector` module.
Changes include:
* Grouping methods of similar purpose into separate files
* Putting associated tests in the same file as the implementation
* Adding or expanding documentation on existing methods
* Normalizing documentation patterns and format.
