--- src/vector/ops/geometry/mod.rs
@@ -1,1 +1,0 @@
-pub mod intersection;
